Decompile func0f148b40

+++ src/include/game/data/data_020df0.h
@@ -184,7 +184,7 @@ extern u32 var8007f75c;
 extern u32 var8007f7b0;
 extern s32 var8007f840;
 extern u8 var8007f844;
-extern u32 var8007f848;
+extern u8 var8007f848;
 extern u32 var8007f860;
 extern u32 var8007f864;
 extern u32 var8007f868;
